refactor(invitation): standardize DTO structure and domain types (Phase 1d)
- Move Invitation DTOs from invitationService to types/invitation/
- Create GenerateInviteCodeRequest and ValidateInviteCodeRequest DTOs
- Create InvitationDraft type for pre-submission preview objects
- Update invitationService and mutation hooks to import from domain layer
