+++ types/invitation/invitation-validate.ts
@@ -1,0 +1,11 @@
+import { Invitation } from './invitation';
+
+export interface ValidateInviteCodeRequest {
+  code: string;
+}
+
+export interface ValidateInviteCodeResponse {
+  valid: boolean;
+  invitation?: Invitation;
+  message?: string;
+}
